pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Вычислитель отличий"
authors = ["Shaydullin Ilfat <sweden.fan@Yandex.ru>"]
readme = "README.md"

packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.main:main"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
coverage = "^7.6.4"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
